requirements-webapp.txt
fastapi
uvicorn
python-multipart
requests
